Include `offset` in return value of `#records`
Currently, there are 2 modes for retrieving records; pagination on or
off.

If pagination is on (`paginate: true`), we keep making fetch
requests until the response does not include an `offset` for the next
page.

If pagination is off (`paginate: false`), we simply fetch the first page
and stop there.

In order to allow users to control pagination themselves (rather than
simply on or off, as above) we need to provide access to the `offset`
value returned by Airtable.